+++ .github/workflows/test_events/workflow_run.json
@@ -1,0 +1,5 @@
+{
+    "workflow_run": {
+        "conclusion": "success"
+    }
+  }
